feat: integer world-market prices and catalog fallback (Refs #3093) (#3098)
* feat: integer world-market prices and catalog fallback (Refs #3093)

Slice 1 of #3093 — foundational price model and display:

- WorldMarketState.prices is Map<CommodityId, int>; fromJson floors legacy
  doubles and clamps negatives to 0 for backward-compatible saves.
- world_market_phase floors PriceDiscovery results at persistence boundary;
  DealMatcher still receives double prices at the call site.
- ResourceRules.defaultMarketPriceForCommodityId exposes catalog defaults
  for raw-resource commodities.
- Trade screen Market tab renders integer prices with catalog fallback
  for unseeded raw resources; manufactured commodities still show em-dash
  until in-game price discovery (follow-up slice).
- SPEC/game/world-market.md and SPEC/ui/trade-screen.md updated.
- Tests across models, data, logic, AI, and app packages.


* feat(trade): sellable readout + offer-side clamp on Market tab (Refs #3093)

- packages/colonizethis_logic: add `offerCapByCommodityId`,
  `stagedOfferQuantitiesByCommodityId`, and `sellableHeadroomByCommodityId`
  helpers under `economy/world_market/sellable_quantity.dart` and export
  them from `colonizethis_logic.dart`. Riches commodities are excluded.
  Production-input projection is a planned refinement (current slice
  treats `offerCap == raw stockpile`); the helper signature already takes
  `Game` + `playerId` so wiring projection in is backwards-compatible.
- app/lib/features/game/screens/trade_screen.dart: each Market row now
  renders a `(N)` sellable readout (`marketRowSellableReadoutKey`),
  gates the Offer chip and offer-side `+` against the per-commodity
  offer cap, and silently clamps the staged quantity on direction toggle.
  Bid-side cargo cap from E5c is unchanged.
- app/lib/widgets/ct_choice_chip.dart: accept a nullable `onSelected`
  callback and render a disabled visual state (muted background, border,
  and label) so the Trade screen can communicate "no offer possible"
  without unmounting the chip.
- Tests: add `world_market_sellable_quantity_test.dart` (logic helpers,
  13 cases) and `trade_screen_market_tab_sellable_clamp_test.dart`
  (widget AC coverage incl. zero-stockpile disabling, increment
  saturation, toggle clamp, and bid/offer isolation). Update existing
  E5b + E8 fixtures to seed baseline stockpiles so legacy ACs continue
  to pass under the new offer clamp.
- SPEC/ui/trade-screen.md + SPEC/game/world-market.md: document the
  sellable readout, the offer-side clamp behavior, AC coverage, and
  the canonical logic-side helpers.


* fix: ct_repo_lint compliance for Refs #3093 slice 2

- Switch `world_market_sellable_quantity_test.dart` to the
  `package:colonizethis_test/test.dart` entrypoint required by
  `[repo.test_imports]` per `SPEC/program/test-logging.md`.
- Split the Deal Book widgets out of `trade_screen.dart` into a
  sibling `trade_screen_deal_book.dart` part fragment so the host
  file drops back under the 1000 non-comment line cap enforced by
  `[repo.dart_file_non_comment_line_size]`
  (`SPEC/program/dart-file-non-comment-line-size.md`). All
  `_DealBook*` classes stay library-private via `part of` and
  continue to reference the same `TradeScreen.dealBook*` constants
  unchanged.

Refs #3093


---------

+++ packages/colonizethis_ai/lib/src/planning/treasury_planner.dart
@@ -290,7 +290,7 @@ Map<CommodityId, int> _inputNeedsFromAssignments(
 
 bool _marketPriceBelowProductionCost(
   CommodityId commodityId,
-  Map<CommodityId, double> marketPrices,
+  Map<CommodityId, int> marketPrices,
 ) {
   final marketPrice = marketPrices[commodityId];
   if (marketPrice == null) return true;
@@ -299,7 +299,7 @@ bool _marketPriceBelowProductionCost(
     if (recipe.outputCommodityId != commodityId) continue;
     var inputCost = 0.0;
     for (final entry in recipe.inputQuantities.entries) {
-      final inputPrice = marketPrices[entry.key] ?? 0.0;
+      final inputPrice = marketPrices[entry.key] ?? 0;
       inputCost += inputPrice * entry.value;
     }
     final perUnit = inputCost / recipe.outputQuantity;
@@ -369,7 +369,7 @@ Map<CommodityId, int> _carryForwardQuantitiesByCommodity({
 /// nominal sum (full-fill credit) per the F2/F3 default. Refs #2994 F8.
 int _expectedOfferInflow({
   required Map<CommodityId, int> available,
-  required Map<CommodityId, double> marketPrices,
+  required Map<CommodityId, int> marketPrices,
   required WorldMarketState state,
 }) {
   if (available.isEmpty) return 0;
@@ -378,8 +378,8 @@ int _expectedOfferInflow({
     final commodityId = entry.key;
     final quantity = entry.value;
     if (quantity <= 0) continue;
-    final price = marketPrices[commodityId] ?? 0.0;
-    if (price <= 0.0) continue;
+    final price = marketPrices[commodityId] ?? 0;
+    if (price <= 0) continue;
     final fillRate = _priorTurnOfferFillRate(state, commodityId);
     inflow += quantity * price * fillRate;
   }
